requirements.txt
pandas==2.2.1
numpy==1.26.2
matplotlib==3.8.3
opencv-python==4.9.0.80
Pillow==9.5.0